refactor : point 충전, 사용 response 변경

+++ src/main/java/kr/hhplus/be/server/interfaces/dto/UserDTO.java
@@ -2,17 +2,15 @@ package kr.hhplus.be.server.interfaces.dto;
 
 import kr.hhplus.be.server.domain.point.Point;
 
-import java.math.BigDecimal;
-
 public class UserDTO {
 	public record UserResponse(
 		Long id,
-		BigDecimal point
+		Integer point
 	){
 		public static UserResponse from(Point point) {
 			return new UserResponse(
 				point.getUserId(),
-				point.getBalance()
+				point.getBalance().intValue()
 			);
 		}
 	}
